Offload TypeScript compiler to Worker thread
